Adding support for a provider-level compile step that knows about all products
