+++ compiler.rb
@@ -34,7 +34,7 @@ require 'provider/terraform_oics'
 require 'provider/terraform_object_library'
 require 'pp' if ENV['COMPILER_DEBUG']
 
-product_names = nil
+products_to_compile = nil
 all_products = false
 yaml_dump = false
 output_path = nil
@@ -50,7 +50,7 @@ Google::LOGGER.level = Logger::INFO
 # rubocop:disable Metrics/BlockLength
 OptionParser.new do |opt|
   opt.on('-p', '--product PRODUCT', Array, 'Folder[,Folder...] with product catalog') do |p|
-    product_names = p
+    products_to_compile = p
   end
   opt.on('-a', '--all', 'Build all products. Cannot be used with --product.') do
     all_products = true
@@ -86,32 +86,35 @@ OptionParser.new do |opt|
 end.parse!
 # rubocop:enable Metrics/BlockLength
 
-raise 'Cannot use -p/--products and -a/--all simultaneously' if product_names && all_products
-raise 'Either -p/--products OR -a/--all must be present' if product_names.nil? && !all_products
+raise 'Cannot use -p/--products and -a/--all simultaneously' if products_to_compile && all_products
+raise 'Either -p/--products OR -a/--all must be present' if products_to_compile.nil? && !all_products
 raise 'Option -o/--output is a required parameter' if output_path.nil?
 raise 'Option -e/--engine is a required parameter' if provider_name.nil?
 
+all_product_files = []
+Dir['products/**/api.yaml'].each do |file_path|
+  all_product_files.push(File.dirname(file_path))
+end
+
+if override_dir
+  Dir["#{override_dir}/products/**/api.yaml"].each do |file_path|
+    product = File.dirname(Pathname.new(file_path).relative_path_from(override_dir))
+    all_product_files.push(product) unless products_to_compile.include? product
+  end
+end
+
+
 if all_products
-  product_names = []
-  Dir['products/**/api.yaml'].each do |file_path|
-    product_names.push(File.dirname(file_path))
-  end
-
-  if override_dir
-    Dir["#{override_dir}/products/**/api.yaml"].each do |file_path|
-      product = File.dirname(Pathname.new(file_path).relative_path_from(override_dir))
-      product_names.push(product) unless product_names.include? product
-    end
-  end
-
-  raise 'No api.yaml files found. Check provider/engine name.' if product_names.empty?
+  products_to_compile = all_product_files
+  raise 'No api.yaml files found. Check provider/engine name.' if products_to_compile.empty?
 end
 
 start_time = Time.now
 
+products_for_version = []
 provider = nil
 # rubocop:disable Metrics/BlockLength
-product_names.each do |product_name|
+all_product_files.each do |product_name|
   product_override_path = ''
   provider_override_path = ''
   product_override_path = File.join(override_dir, product_name, 'api.yaml') if override_dir
@@ -143,9 +146,7 @@ product_names.each do |product_name|
   raise "Output path '#{output_path}' does not exist or is not a directory" \
     unless Dir.exist?(output_path)
 
-  Google::LOGGER.info "Compiling '#{product_name}' (at #{version}) output to '#{output_path}'"
-  Google::LOGGER.info \
-    "Generating types: #{types_to_generate.empty? ? 'ALL' : types_to_generate}"
+  Google::LOGGER.info "Loading '#{product_name}' (at #{version})'"
 
   product_api = Api::Compiler.new(product_yaml).run
   product_api.validate
@@ -166,6 +167,17 @@ product_names.each do |product_name|
     product_api, provider_config, = \
       Provider::Config.parse(provider_override_path, product_api, version, override_dir)
   end
+  products_for_version.push(product_api.name)
+
+  unless products_to_compile.include?(product_name)
+    Google::LOGGER.info "Skipping product '#{product_name}' as it was not specified to be compiled"
+    next
+  end
+
+  Google::LOGGER.info "Compiling '#{product_name}' (at #{version}) output to '#{output_path}'"
+  Google::LOGGER.info \
+    "Generating types: #{types_to_generate.empty? ? 'ALL' : types_to_generate}"
+
 
   pp provider_config if ENV['COMPILER_DEBUG']
 
@@ -194,5 +206,6 @@ end
 # of the loop
 provider&.copy_common_files(output_path, version)
 provider&.compile_common_files(output_path, version)
+provider&.compile_provider_files(output_path, products_for_version, version)
 
 # rubocop:enable Metrics/BlockLength
